requirements.txt
streamlit
streamlit-drawable-canvas==0.8.0
ruff==0.9.1
streamlit_javascript
pillow
